# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: MasterAnnonce API
  version: 1.0.0
  description: API REST securisee avec authentification stateless via JAAS + Bearer token.
servers:
  - url: /api
tags:
  - name: Auth
  - name: Annonces
paths:
  /login:
    post:
      tags: [Auth]
      summary: Authentifier un utilisateur et retourner un token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
      responses:
        '200':
          description: Authentification reussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /annonces:
    get:
      tags: [Annonces]
      summary: Lister les annonces (pagination + filtres)
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Liste paginee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAnnonceResponse'
        '400':
          description: Parametres invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
    post:
      tags: [Annonces]
      summary: Creer une annonce (auteur derive du token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnonceCreateDTO'
      responses:
        '201':
          description: Annonce creee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnonceResponseDTO'
        '400':
          description: Requete invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '401':
          description: Non authentifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /annonces/{id}:
    get:
      tags: [Annonces]
      summary: Recuperer une annonce par id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detail annonce
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnonceResponseDTO'
        '404':
          description: Non trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
    put:
      tags: [Annonces]
      summary: Mettre a jour une annonce
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnonceUpdateDTO'
      responses:
        '200':
          description: Annonce modifiee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnonceResponseDTO'
        '401':
          description: Non authentifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '403':
          description: Interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '404':
          description: Non trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '409':
          description: Conflit metier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
    patch:
      tags: [Annonces]
      summary: Changer le statut d'une annonce
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnonceStatusDTO'
      responses:
        '200':
          description: Statut modifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnonceResponseDTO'
        '401':
          description: Non authentifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '403':
          description: Interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '404':
          description: Non trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '409':
          description: Transition invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
    delete:
      tags: [Annonces]
      summary: Supprimer une annonce archivee
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Supprime
        '401':
          description: Non authentifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '403':
          description: Interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '404':
          description: Non trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
        '409':
          description: Archivage requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDTO'
  /openapi:
    get:
      tags: [Auth]
      summary: Recuperer la specification OpenAPI YAML
      responses:
        '200':
          description: Specification OpenAPI
          content:
            text/plain:
              schema:
                type: string
            application/yaml:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  schemas:
    LoginDTO:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponseDTO:
      type: object
      properties:
        token:
          type: string
        expiresIn:
          type: integer
          format: int64
    AnnonceCreateDTO:
      type: object
      required: [title, description, adress, mail, categoryId]
      properties:
        title:
          type: string
          maxLength: 64
        description:
          type: string
          maxLength: 256
        adress:
          type: string
          maxLength: 64
        mail:
          type: string
          format: email
        categoryId:
          type: integer
          format: int64
    AnnonceUpdateDTO:
      type: object
      required: [title, description, adress, mail, categoryId]
      properties:
        title:
          type: string
        description:
          type: string
        adress:
          type: string
        mail:
          type: string
          format: email
        categoryId:
          type: integer
          format: int64
    AnnonceStatusDTO:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [publish, archive]
    ApiErrorDTO:
      type: object
      properties:
        error:
          type: string
        messages:
          type: array
          items:
            type: string
    AuthorDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
    AnnonceResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        adress:
          type: string
        mail:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: string
        author:
          $ref: '#/components/schemas/AuthorDTO'
        category:
          $ref: '#/components/schemas/CategoryDTO'
    PaginatedAnnonceResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnnonceResponseDTO'
        totalCount:
          type: integer
          format: int64
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer